requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# MongoDB
pymongo==4.10.1

# Environment Variables
python-dotenv==1.0.1

# AWS Bedrock
boto3>=1.35.0
botocore>=1.35.0

# Azure Functions
azure-functions>=1.18.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.1
